requirements.txt
# albumentations
hydra-core==1.1.2
# ivtmetrics - check compatibility or find alternative
matplotlib>=3.5.1
neptune-client>=0.14.2
neptune-notebooks>=0.9.0
numpy>=1.19.5
omegaconf>=2.1.2
opencv-python # specify version if needed
pandas>=1.4.4
scikit-learn>=0.24.2
scipy>=1.7.1
seaborn>=0.11.2
timm>=0.6.5
torch==2.0.1
torchvision==0.15.2
